requirements.txt
amqp==2.6.0
asgiref==3.2.7
attrs==19.3.0
billiard==3.6.3.0
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
entrypoints==0.3
gunicorn==20.0.4
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.2
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
django-embed-video==1.3.3

